Turn AbstractMatch into a configurable superclass
Absorb the functionalities of VariableTarget and ObtainInfoMatch to
avoid a deep and complex class hierarchy
- Use PROPERTY_USAGE_GROUP to create dynamic groups in the Inspector,
which can be added or removed by setting the same name variable to true
or false, respectively
- Implement _get_property_list() to dynamically change the properties
shown in the Inspector, always remembering to notify when the property
list changed

Create auxiliary functions to setup(), so that subclasses dont need to
override this method. This hides the complexity from users that want to
create new matches

Expand Template Method on is_satisfied() to accomodate when some data
is extracted or retrieved from the target node
- When Data_Extraction = true, there is no need to implement _get_data(),
since one of the premade options will be used
- When Data_Retrieval = true, _get_data() needs to return the data that
will be added to the bindings

Add the option to retrieve data, for when a match is valid in an interval
and we want to use the exact value achieved somewhere. This is done by
defining a variable that can be accessed with the '?' prefix

Automate JSON-format string generation and self building
- Use new variables match_id and repr_vars, that must be initialized in
every class

Signed-off-by: Rodrigo Volpe Battistin <rv.battistin@usp.br>