Centralize RuleFactory on RuleList
Because RuleFactory searches through all classes registered in the
project on initialization, creating multiple instances would be costly.
Since RuleList is the highest access class that requires a factory,
and they are needed in loading a resource from a file, for example,
we need to ensure that creating a new() RuleList comes with
an instance of RuleFactory. Then, the rules and components below it
can access the same object via reference.

Signed-off-by: Rodrigo Volpe Battistin <rv.battistin@usp.br>