Reorganize DatumMatch preconfigs hierarchy
Configuration: functions needed to be implemented, optional*
---
- nothing: is_satisfied(bindings)
- Tester_Node: _node_satisfies_match(node, bindings), _get_candidates()*
  - Data_Based_Node: _get_data(node), _data_satisfies_match(data)
    - Get_Node_Data_Preset: _data_satisfies_match(data)

It didn't make sense to be able to retrieve some data from a node and
have the match not be linked to said data. The Data_Extraction group
was actually just some presets for _get_data, so it should depend on
wether the match is based on data or not. There was no need to have a
property group just for Data_Retrieval

Signed-off-by: Rodrigo Volpe Battistin <rv.battistin@usp.br>