Add match to see if variables have distinct values
There are cases where the same object is a valid substitution for two
separate variables, but we want to make sure they can be applied
with distinct values. DistinctVariablesMatch makes sure that there is a
valid substitution that has distinct values for all selected variables

Create and execute a test for this new atomic match. Success

Signed-off-by: Rodrigo Volpe Battistin <rv.battistin@usp.br>